feat(referral): open referral-program as a popup modal from the banner
Tapping the internal referral banner now shows the referral-program page in a
popup dialog instead of navigating to /referral-program.

- Extract the page body into a reusable ReferralProgramContent component with
  an optional onClose (modal dismiss vs. route back/replace).
- Add ReferralProgramModal, a controlled Dialog reusing the existing dialog
  primitive with a height-bounded ScrollView.
- Refactor the /referral-program route to render the shared content (kept for
  deep links and fallback navigation).
- Wire the banner to open the modal instead of routing.
